// src/models/cardOptions.ts
type cardOptions =
  | "A♥"
  | "2♥"
  | "3♥"
  | "4♥"
  | "5♥"
  | "6♥"
  | "7♥"
  | "8♥"
  | "9♥"
  | "10♥"
  | "J♥"
  | "Q♥"
  | "K♥"
  | "A♣"
  | "2♣"
  | "3♣"
  | "4♣"
  | "5♣"
  | "6♣"
  | "7♣"
  | "8♣"
  | "9♣"
  | "10♣"
  | "J♣"
  | "Q♣"
  | "K♣"
  | "A♠"
  | "2♠"
  | "3♠"
  | "4♠"
  | "5♠"
  | "6♠"
  | "7♠"
  | "8♠"
  | "9♠"
  | "10♠"
  | "J♠"
  | "Q♠"
  | "K♠"
  | "A♦"
  | "2♦"
  | "3♦"
  | "4♦"
  | "5♦"
  | "6♦"
  | "7♦"
  | "8♦"
  | "9♦"
  | "10♦"
  | "J♦"
  | "Q♦"
  | "K♦";

export default cardOptions;